Task 2: Changing component styles by clicking checkbox
