+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\index.js":"1","D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\App.js":"2","D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Header\\Header.js":"3","D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Card\\Card.js":"4"},{"size":151,"mtime":1610461001081,"results":"5","hashOfConfig":"6"},{"size":396,"mtime":1610548064656,"results":"7","hashOfConfig":"6"},{"size":217,"mtime":1610541221372,"results":"8","hashOfConfig":"6"},{"size":790,"mtime":1610548063581,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"1571d7p",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\index.js",[],[],"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\App.js",[],"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Header\\Header.js",[],"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Card\\Card.js",[]]
+[{"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\index.js":"1","D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Card\\index.js":"2","D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Header\\index.js":"3","D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Card\\Card.jsx":"4","D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Header\\Header.jsx":"5","D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\App.js":"6"},{"size":151,"mtime":1610623637560,"results":"7","hashOfConfig":"8"},{"size":49,"mtime":1610623754619,"results":"9","hashOfConfig":"8"},{"size":55,"mtime":1610623600187,"results":"10","hashOfConfig":"8"},{"size":780,"mtime":1610624021795,"results":"11","hashOfConfig":"8"},{"size":146,"mtime":1610623744016,"results":"12","hashOfConfig":"8"},{"size":386,"mtime":1610623714825,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1571d7p",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\index.js",[],"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Card\\index.js",[],"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Header\\index.js",[],[],"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Card\\Card.jsx",[],"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\Header\\Header.jsx",[],"D:\\MHaurylenka\\Courses\\react.js\\react-js-app\\src\\App.js",[]]
