+++ src/Header/Header.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import './Header.css';
+
+const Header = () => <header className="header">Header</header>;
+
+export default Header;
